// src/generated/models/OpdWithOperationRowDTO.ts
// tslint:disable
/**
 * OH 2.0 Api Documentation
 * OH 2.0 Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { OpdDTO, OperationRowDTO } from "./";

/**
 * @export
 * @interface OpdWithOperationRowDTO
 */
export interface OpdWithOperationRowDTO {
  /**
   * @type {OpdDTO}
   * @memberof OpdWithOperationRowDTO
   */
  opdDTO?: OpdDTO;
  /**
   * @type {Array<OperationRowDTO>}
   * @memberof OpdWithOperationRowDTO
   */
  operationRows?: Array<OperationRowDTO>;
}